Transparancy update
I added some transparency to the background, and a margin around the images so that you have a better view of the field while looking at the Selection Window. As usual, the height and width are now definite numbers found testing with every resolution (the bottom of the Selection Window practically lines up with the bottom of the Duel Field in 1152x864). The width was slightly reduced, since the edge of the Duel Field is not the edge of the screen. As mentioned earlier, the border matches the colour of the scrollbar. There's now a margin on top, once again to make the Selection Window appear relatively "centered". I indicated to our Mac users that the Selection Window is scrollable by adding cursor: s-resize (since the proper "scroll" cursor indicates all direction, and looks a little clunky in my opinion). Finally, I identified the z-index for future projects.