DNA continuous traps fix
A bug was reported today, so I fixed it. In this bug, a list of check boxes ran off the screen when using DNA Surgery. This fix applies a scrollbar to such effects as DNA surgery and DNA Transplant (as well as possibly DNA checkup and Homunculus, which remain untested). This fix should suffice for the time being, however, in a future update, I may arrange the check boxes into rows of 2, if I'm up to the challenge.

I propose we add this to our CSS scrolling fixes, as it does use a scrollbar. I will most likely be modifying it to apply to v1.4 of Jaden and Yubel.