Update
The chat box will no longer move the search bar in deck constructor. It's also been moved up so as to not trigger appear-on-hover taskbars. The width of the entire column has been matched and will be tested to make sure that it doesn't kick the duel off-screen in any resolution.