Tested for recent update.
Namespace was removed by stylus. I'm not sure if it's needed.
I also adjusted the font size and #card-description-box so as to not resize while in full screen or be unreadable at low resolutions. I noticed that on some resolutions long usernames broke into the #card-column, so I applied a word break.
Should I add Namespace back in or leave it? @Skeletonxf