Create Hold still while I'm tryna click you!
This small code flips the table upside down to increase the chance of a successful click. In exchange for a need to scroll, the duels should remain still for longer, as duels are usually being opened more often than they are being closed. Apparently there's a way to apply options to a style, allowing users to download its code with certain options on or off. I may use similar code to test this on my main theme, if possible. After all, Version 2 should be a big deal.